[hardpy, utils] Add DUT dataclass
