+++ hardpy/pytest_hardpy/utils/__init__.py
@@ -21,13 +21,14 @@ from hardpy.pytest_hardpy.utils.exception import (
     TestStandNumberError,
     WidgetInfoError,
 )
-from hardpy.pytest_hardpy.utils.instrument import Instrument
 from hardpy.pytest_hardpy.utils.machineid import machine_id
 from hardpy.pytest_hardpy.utils.node_info import NodeInfo
 from hardpy.pytest_hardpy.utils.progress_calculator import ProgressCalculator
 from hardpy.pytest_hardpy.utils.singleton import SingletonMeta
+from hardpy.pytest_hardpy.utils.user_type import DUT, Instrument
 
 __all__ = [
+    "DUT",
     "BaseWidget",
     "CheckboxWidget",
     "CompOp",
